+++ longestConsecutiveSequence.cpp
@@ -32,7 +32,7 @@ public:
 
         int currentLongest = 1, longest = 0;
 
-        for(int i = 0; i < nums.size() - 1; i++)
+        for(size_t i = 0; i < nums.size() - 1; i++)
         {
             if(nums[i] != nums[i + 1])
             {
